--- src/modules/film-genre/chart.ts
@@ -1,3 +1,0 @@
-import { mainChart } from "./app";
-
-void mainChart();
