--- src/modules/film-genre/film-genre.ts
@@ -1,3 +1,0 @@
-import { mainFilmGenre } from "./app";
-
-void mainFilmGenre();
