--- src/modules/genre_pages/main.ts
@@ -1,3 +1,0 @@
-import { main } from "./app";
-
-void main();
